React route renders correctly on /react path. This is a temp path
