+++ public/css/app.css
@@ -5,7 +5,7 @@
 #home > .title-container {
   width: 100%;
   background: #48be34;
-  padding-top: 100px;
+  padding-top: 50px;
   padding-bottom: 50px;
   text-align: center;
 }
