+++ resources/assets/js/components/App.js
@@ -1,0 +1,28 @@
+import React, { Component } from 'react';
+import ReactDOM from 'react-dom';
+
+export default class Example extends Component {
+    render() {
+        return (
+            <div class='content'>
+                <div className="container">
+                    <div className="row">
+                        <div className="col-md-8 col-md-offset-2">
+                            <div className="panel panel-default">
+                                <div className="panel-heading">Example Component</div>
+
+                                <div className="panel-body">
+                                    I'm an example component!
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        );
+    }
+}
+
+if (document.getElementById('react')) {
+    ReactDOM.render(<Example />, document.getElementById('react'));
+}
