+++ resources/assets/sass/_home.scss
@@ -7,7 +7,7 @@
         background: rgb(72, 190, 52);
         // height: 300px;
         padding: {
-            top: 100px;
+            top: 50px;
             bottom: 50px;
         }
         text-align: center;
